Add some dashboard tests
